Started refactoring the code.
Use guard clause to promise “Sulfuras” to be quality 80 at all times.
            if item.name == "Sulfuras, Hand of Ragnaros":
                item.quality = 80
                continue
Including the fact that Sulfuras doenst have the sell_in decrease as it will not be sold at any time.
So no need to decrease sell_in.

Remove the negative expression as now we handle the case where the item does equal Sulfuras, Hand of Ragnaros.
if item.name != "Sulfuras, Hand of Ragnaros":

#note -> we still use the below to maintain the logic for all other items -> where sell_in date decreases by one at each passing day.